services/classification/pyproject.toml
[project]
name = "bloasis-classification"
version = "0.1.0"
description = "BLOASIS Classification Service - Stock Selection Pipeline Stage 1-2"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.59.0",
    "grpcio-tools>=1.59.0",
    "grpcio-health-checking>=1.59.0",
    "googleapis-common-protos>=1.61.0",
    "redis[hiredis]>=5.0.1",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["/workspace", "/workspace/shared/generated"]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "src/main.py"]

[tool.coverage.report]
fail_under = 80
